Initial implementation of core.base_expr class (#1466)

+++ c/expr/py_expr.h
@@ -59,7 +59,7 @@ typedef void (*mapperfn)(int64_t row0, int64_t row1, void** params);
 typedef void (*gmapperfn)(const int32_t* groups, int32_t grp, void** params);
 
 Column* unaryop(int opcode, Column* arg);
-Column* binaryop(int opcode, Column* lhs, Column* rhs);
+Column* binaryop(size_t opcode, Column* lhs, Column* rhs);
 Column* reduceop(int opcode, Column* arg, const Groupby& groupby);
 Column* reduce_first(const Column* col, const Groupby& groupby);
